+++ backend/canisters/user/api/src/updates/end_video_call.rs
@@ -2,7 +2,7 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{MessageId, UserId};
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
 pub struct Args {
     pub user_id: UserId,
     pub message_id: MessageId,