+++ backend/canisters/user/api/src/updates/start_video_call.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{MessageId, UserId};
+use types::{MessageId, Milliseconds, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,6 +9,7 @@ pub struct Args {
     pub initiator_username: String,
     pub initiator_display_name: Option<String>,
     pub initiator_avatar_id: Option<u128>,
+    pub max_duration: Option<Milliseconds>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]