+++ backend/canisters/user/impl/src/updates/end_video_call.rs
@@ -1,4 +1,5 @@
 use crate::guards::caller_is_video_call_operator;
+use crate::timer_job_types::TimerJob;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_tracing_macros::trace;
 use chat_events::{EndVideoCallResult, Reader};
@@ -10,10 +11,21 @@ use user_canister::end_video_call::{Response::*, *};
 fn end_video_call(args: Args) -> Response {
     run_regular_jobs();
 
-    mutate_state(|state| end_video_call_impl(args, state))
+    mutate_state(|state| {
+        state.data.timer_jobs.cancel_job(
+            |job| {
+                if let TimerJob::MarkVideoCallEnded(vc) = job {
+                    vc.0 == args
+                } else {
+                    false
+                }
+            },
+        );
+        end_video_call_impl(args, state)
+    })
 }
 
-fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> Response {
+pub(crate) fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> Response {
     if let Some(chat) = state.data.direct_chats.get_mut(&args.user_id.into()) {
         let was_started_by_me = chat
             .events