+++ backend/integration_tests/src/video_call_tests.rs
@@ -3,10 +3,14 @@ use crate::rng::{random_message_id, random_string};
 use crate::utils::tick_many;
 use crate::{client, TestEnv};
 use std::ops::Deref;
+use std::time::Duration;
+use test_case::test_case;
 use types::{ChatEvent, MessageContent, VideoCallContent};
+use utils::time::HOUR_IN_MS;
 
-#[test]
-fn start_join_end_video_call_in_direct_chat_succeeds() {
+#[test_case(true)]
+#[test_case(false)]
+fn start_join_end_video_call_in_direct_chat_succeeds(manually_end_video_call: bool) {
     let mut wrapper = ENV.deref().get();
     let TestEnv {
         env,
@@ -18,8 +22,9 @@ fn start_join_end_video_call_in_direct_chat_succeeds() {
     let user1 = client::register_diamond_user(env, canister_ids, *controller);
     let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
     let message_id = random_message_id();
+    let max_duration = HOUR_IN_MS;
 
-    client::user::happy_path::start_video_call(env, &user1, user2.user_id, message_id);
+    client::user::happy_path::start_video_call(env, &user1, user2.user_id, message_id, Some(max_duration));
 
     tick_many(env, 3);
 
@@ -71,8 +76,13 @@ fn start_join_end_video_call_in_direct_chat_succeeds() {
         .event;
     assert_is_video_message(chat2_event, |v| v.participants.len() == 2);
 
-    client::user::happy_path::end_video_call(env, user1.user_id, user2.user_id, message_id);
-    client::user::happy_path::end_video_call(env, user2.user_id, user1.user_id, message_id);
+    if manually_end_video_call {
+        client::user::happy_path::end_video_call(env, user1.user_id, user2.user_id, message_id);
+        client::user::happy_path::end_video_call(env, user2.user_id, user1.user_id, message_id);
+    } else {
+        env.advance_time(Duration::from_millis(max_duration));
+        env.tick();
+    }
 
     let user1_chat = client::user::happy_path::initial_state(env, &user1)
         .direct_chats
@@ -91,8 +101,9 @@ fn start_join_end_video_call_in_direct_chat_succeeds() {
     assert!(user2_chat.video_call_in_progress.is_none());
 }
 
-#[test]
-fn start_join_end_video_call_in_group_chat_succeeds() {
+#[test_case(true)]
+#[test_case(false)]
+fn start_join_end_video_call_in_group_chat_succeeds(manually_end_video_call: bool) {
     let mut wrapper = ENV.deref().get();
     let TestEnv {
         env,
@@ -107,8 +118,9 @@ fn start_join_end_video_call_in_group_chat_succeeds() {
     client::local_user_index::happy_path::join_group(env, user2.principal, canister_ids.local_user_index, group);
 
     let message_id = random_message_id();
+    let max_duration = HOUR_IN_MS;
 
-    client::group::happy_path::start_video_call(env, &user1, group, message_id);
+    client::group::happy_path::start_video_call(env, &user1, group, message_id, Some(max_duration));
 
     let summary = client::group::happy_path::summary(env, &user1, group);
     assert!(summary.video_call_in_progress.is_some());
@@ -129,7 +141,12 @@ fn start_join_end_video_call_in_group_chat_succeeds() {
         .event;
     assert_is_video_message(event, |v| v.participants.len() == 2);
 
-    client::group::happy_path::end_video_call(env, group, message_id);
+    if manually_end_video_call {
+        client::group::happy_path::end_video_call(env, group, message_id);
+    } else {
+        env.advance_time(Duration::from_millis(max_duration));
+        env.tick();
+    }
 
     let summary = client::group::happy_path::summary(env, &user1, group);
     assert!(summary.video_call_in_progress.is_none());